First draft of master unit list.
I've pulled much of the functionality out of the notebook and into two 
modules. `plant_parts.py` is mostly just the inputs required to 
aggregate each of these plant parts. `plant_part_agg_eia.py` holds 
mostly contains an object with a set of methods that preform the 
aggregations.

Closes Issue #3. Ownership. Because ownership is associated with 
generators, I rearranged the aggregations. Now, we are aggregating all 
of the fields for the genetors, slicing that out for each ownership 
slice (and a fabricated total), and then aggregating on the rest of the 
plant parts.

Working on Issue #4. Weighted Averages. Instead of using the standard 
'mean' function, I've converted the aggregations to take custom weighted 
averages based on capacity or net generation. Right now, only the net 
generation works (I'm getting an error on Zeros when using net gen).

Closes Issue #1. Finishing definitions for each plant part. I think this 
is complete! The one data point that is not included right now is 
`fuel_consumed_mmbtu`, which is originally reported at the 
plant/fuel_type/prime_mover level. Because it does not clearly 
disaggregate to generators, there isn't a clean way to do this with the 
current process. We could certainly come up with something if we think 
this is required.