Update master unit list compilation.
I added in the utility_id_eia column in the aggreations! which is 
important, turns out. Now there are much much more records in general. I 
also cleaned up the capacity column to ensure that basically all of 
records now have capacity and net gen. There is still a big gap in data 
on the fuel cost and heat rate. I also re-made the weighted average to 
make it more straightward and less error prone.