Updated master unit list.. again!
I did a few things to ensure we got more records and cleaned frequency 
aggregations and the merges.

On ownership: a ton of generators were not in the original ownership 
table. I basically assumed that if a generator is not in the ownership 
table, it is 100% owned by it's operator.

On cleaning: mostly there were data type issues and aggregations being 
done with the wrong merges. I streamlined the aggregations so they all 
run through one function `agg_cols`. I also cleaned up how we were 
aggregating annually, that the weighted averages were sensitive to 
datetime and ensured that we were using the weighted averages at each 
aggregation step.

I also added a slew of new plots into the notebook. I did not clear the 
notebook results so you can look at the plots without running the 
notebook (in case y'all were having trouble getting the devlopment 
environment up and running).