Clean up ownership aggregations.
Previously I was using the utiltiy_id col from the ownership_eia860 
table as the owner instead of the owner_utility_id. I changed the column 
being pulled and used, but I ended up renaming the owner_utility_id_eia 
to be utility_id_eia to conform with how it is reported in the rest of 
the tables.  Thanks to Jon for leading me to this one. After some more 
spot checking, I've added some assertions (things that will cause 
errors).