Minor data type change.
We changed some of the default data types in the main pudl repo so I had 
to changed a few data types here. I aslo generated a little process that 
pretended the operator was the owner.. but I commented this out because 
after developing a better conception of our next stepss, I don't think 
it is the correct approach.