Add qualitaive records and output on df.
added in capacity_factor, utility_id_pudl, plant_id_pudl, 
energy_source_code_1, prime_mover_code, and fuel_type_code_pudl. 

For the latter three (the qualitaitve type records), for each plant 
part, I checked to see if the qualititve records were consistent. If so, 
I associated it, if not, I'm not sure there is a good way to associate. 
For example, a plant that has 3 prime_mover_codes, then there is no true 
pm code for the plant. We could set a consistency threshold (ie a value 
needs to be 75% consistent), but for now I went the simplest and most 
accurate route.

I did this all in a standardized process, so if we want to add other 
columns into the MUL output, it should be relatively simple.