Exploring matching FERC to EIA.
We have matches! They are far from complete and there are tons of things 
left to tweek and explore. To see the process, check out the 
ferc_eia_connections notebook.

There was a far amount of work to standardize ids in the test data, the 
EIA master unit list and the FERC data. There is now a standard FERC and 
EIA id the is generated from the three inputs. Also, for now, I've only 
pulled in the steam table from FERC... we'll add more later.

Once the ids are generated and the columns are standardized, the EIA and 
FERC data re used to create a list of all possible comparisons. Right 
now, I've restricted the comparison by report year and plant_id_pudl. We 
may want to play with removing plant_id_pudl in the future.

Then, the FERC and EIA data are used to create a comparison feature 
vectors with each feature (or element of comparison). Each feature is 
assigned a comparison method resulting in a value between 1 and 0 
depicting the closeness of the possible matches.

Then, the feature vectors and the list of true matches that Sam 
provided, I then run these through a series of logistic regression 
models and test the results. The results are very similar and I believe 
this is in part due to the relatively small quantity of test data. I 
added notes in the top of the notebook about what still needs to be done 
and explored.