removing false granularities in MUL.
Instead of creating a complex process that understands what plant part 
is the "part parent" (or natural larger part that contains the smaller 
part, like a plant to a unit), I basically just assumed the plant itself 
was the parent for all other parts. In this process, I removed every 
records which only had one option for ids in each plant (ie if there was 
one unit id associated with a plant, that unit it really a plant and 
thus we removed the unit record). This cuts the total number of records 
in the MUL down nearly in half and will make the FERC/EIA matching 
easier.

Close Issue #24.