First step in removing peer-peer false granularities.
We needed to remove the plant part records that were more like peer-peer 
granularities (as oppose to parent to child relationships). This commit 
is for posterity. the output here produces a df with many bool columns 
of all of the types of false granularities. Next commit will condense 
all columns into one bool column.

Working on issue #24.