Assigning appropriate ids for false granularities.
Because we know that ~half of the total records in the master unit list 
are functionally the same plant part as one or more different plant 
parts, we needed to label them. In order to work with the training data 
from RMI, we also needed a way to relabel records to their "correct" 
plant part with "correct" id's in order to standardize the two datasets. 
Mostly this was done and stored in CompilePlantParts.part_bools. Closes 
Issue #30.