Add in the plant name to the comparison.
I went with the simplest option here - to add in the name based on the 
"true" part. This is not perfect. If a unit is labeled as a the true 
part, than we are using the unit id not the generator id that is a part 
of that unit.

There are several methods for string comparisions. I landed on 
jarowinkler, mostly because it gave higher average comparison features 
and a normal looking weight for the feature 
(https://recordlinkage.readthedocs.io/en/latest/ref-compare.html#recordlinkage.compare.String).

Issue #32.