Breaking out ownership types for aggregation.
I had previously been treating the aggregations for the ownership slices 
of plants and the 100% ownership slices of plants (faked or real) the 
same. This was wrong! I needed to aggregate the total and owned slices 
slightly differently. So I made a new function (`ag_part`) that does 
just that. Working on Issue #35. I also checked in a new version of the 
master unit list.