Integrate lost fuel records.
In order to get as many fuel cost records as possible, I had to change 
inputs in the pudl mcoe.fuel_cost(). But another chunk of records were 
being lost from the min/max arguments into mcoe. The main change here is 
enabling the min/max limits to be set to none. There is also some 
additional exploration in the mul notebook and different results from 
the ferc/eia mapping.