Add ferc acct MUL type & fuzzy match w/ depretiation
Two main changes here. I've added a ferc acct type as a master unit list 
part granularity. Right now, this is relying on a table in an excel 
sheet. I don't like this - since this table is so small, we should 
probably just store it somewhere.

The other main change was adding a new module which connects the 
depreciation record names with the master unit list records. This is 
then spit out into the same excel workbook as the depreciation data came 
from.

I also incorporated a way to store the master unit list in a compressed 
csv and either access it or generate it.