Switch to true gran in fuzzy matches
Had to pivot to use the true granularies in the fuzzy matches. Mostly 
that entailed renaning columns. Closes Issue #32.

Then there is some light cleaning and honing on assessing the 
connections between the MUL and the depreciation data in 
connect_deprish_to_ferc1.py. This module is still a WIP, but this Closes 
#52.