fix MUL's true_gran labeling
previously, the true_gran labeling was only checking for true/false 
granularities in one direction... the true gran labeling starts with 
counting how many parent parts comprise a plant-part, and if there was 
only one parent per plant-part, then the plant-part was labeled as a 
false gran.

That worked fine for plants and plant_gen's, but it got messier for the 
plant-parts that are not actually always a parent or a child. We needed 
to enforce a heirarchy of plant-parts because not only do we need to 
find the non-unique combination of plant-parts, we need to label only 
one of the non-unique combo as a false granularity. So, checking how 
many sub-parts are inside of a particular plant-part does not always 
give us the right results with this artificial heirarchy.

The fix here was to check how many parent-parts were in a child 
plant-part as well as check how many child-parts are a part of a 
parent-part.

Close Issue #65