fix operational_status and light refactoring.
The operational_status was being added to the master unit list through 
the addition of qualifier records with it's own unique method (RMI wants 
the "most" operational op status). There was a bad merge with the op 
statuses - the df with the "most" op status contained numeric data 
columns as well as utility id and it was supposed to just be the set od 
idx cols for the plant-part. I simply restricted the columns in the df 
with the compiled prioritized op statuses, and now the merge casts the 
op status widely.

I also employed some light refactoring. Mostly to make the MUL 
generation more modular. I converted the references to two of the main 
cached tables to the function which generates them and included a `if 
doesn't table exists or clobber` then make the table setup. This will 
make it easier to call specific functions inside the MUL generation 
without doing a bunch of prep (which only I have a good handle on right 
now).