Label ownership_dupes in MUL, filter in ml matching
The machine learning matches between FERC and EIA were getting a fair 
number of "ties" (in which the score of the top two records were the 
same). This was because there were some duplicates from the MUL where we 
had created the "owned" portion of the generators and a "total" portion 
for all of the utility owners.

The solution was to label the "owned" records which had a 100% ownership 
stake, because those are the records which are duplicated in the "total" 
ownership records. That way we could grab only the true_gran and the 
non-ownership_dupe records in the ml connections. We went from 11% of 
the matches being ties to 1%.