fix fraction_owned aggregation
I've been generating the weighted average of the fraction_owned 
column....
but what I neeeeed to do it find the what portion of the overall 
plant_part
is owned by the utility in question.


For the owned portions I dealt with this via summing the plant's total 
capacity (all owners included)... and then / by the specific record's 
capacity. For the total portions... it should always be 1.

Closes Issue #75