Set standard filepaths. Finalize output object.
The standard input/output file paths are now in init and dispersed 
throught the modules as necessary.

The output object in coordinate now also uses the file paths and were 
generally cleaned up as compared with the original push.

The notebooks were all updated to generate the outputs directly with the 
rmi_out object.

There were also a ton of little tweeks that needed to happen to actually 
make all of these outputs seemlessly talk to eachother. Most were minor 
issues like changing indexes and dropping "_merge" columns.

Working on #128