Mostly updating docs and variable names. Standardize depreciation output so it can be used via the coordinator.
The main depreciation output is aggregated - the input data comes with 
ferc account numbers in each record. This detail is basically never 
needed so the main output is aggregated. The default aggregation process 
was previously not including the utility_id_ferc1. We need this column 
in the deprish_to_eia process so we were recreating it each time in that 
process. I standardized the column outputs so we can use the standard 
depreciation output. This effected both the deprish_to_eia file and the 
coordinate file because now we have a new clobber option.