fix the annual capex additions
the old  version of capex_annual_addition included asset retirement 
costs. these are not included in most accounting of capex additions and 
definitely not included in depreciation studies.

Note: I'd like the cacl_annual_capital_additions_ferc1 to eventually be 
moved into pudl's output layer... and I'm not sure when the full 
capex_total will be the right value to calculate annual additions. 
Perhaps I should add an argument into this method to allow users to pick 
which capex value to use to calculate annual additions.