standardize validation re: agg-ing two dfs
i pulled the funationality of aggregating two dfs to the same index and 
comparing data columns out of the test modules and inot this new 
validate module.

I standardized the agg-ing to accept/compare multiple data columns at 
the same time. I removed on little mask that was only looking at records 
with plant-ids.  this changed a few of the known "baddies".