update known errors for data processing tests
Reasons for updates:
- In last push, I included new depreciation data with updated 
connections between deprish-eia - mostly but not entirely for Mid 
American (utility id 185), so lots of the files remove 185 from the 
known errors. wahoo!
- there are a slew of new additions in the capex_total plant errors. I 
expect this is because the FERC-EIA matching was slightly more 
connection-full with the addition of the operational_year plant-part 
that Kaite added and the more complete opex_nonfuel that Austen added 
into pudl recently
-