Update add_to_training function
First, make sure that overrides to the training data are awknowledged and incorporated in the final training data.

Next, remove any null overrides now listed in the training data. This is because the null overrides are the last thing we do and if there is a record in training and in null overrides it will end up as NA even if it has a match. This fix removes any values from the null overrides list that show up in the training data.